주세은 : 엔진 업데이트

+++ Direct2D/GameEngineCore/GameEngineFolderTexture.h
@@ -24,7 +24,7 @@ public:
 		return Textures.size();
 	}
 
-	GameEngineTexture* GetTexture(int _Index)
+	GameEngineTexture* GetTexture(size_t _Index)
 	{
 		if (Textures.size() <= _Index)
 		{
